backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
langchain-groq
langgraph
langchain
langsmith
bcrypt
python-jose[cryptography]
httpx==0.27.0
slowapi
